Add chainable OCAErrorRecoveryAttempter
You can use this class in together with NSError ans UIAlertView to handle error recovery callbacks, for example to Retry network request. Create RecoveryAttempter, add and connect to the callbacks (recovery options) and then create NSError with this recovery attempter. Objective-Chain does not provide a way to display this error in UIAlertView, yet!